Add files via upload
Project Description
In this project, we’ll explore NASA’s public APIs to retrieve and analyze data related to asteroids. We’ll use the Astronomy Picture of the Day (APOD) endpoint and the Near Earth Object Web Service (NEOWS) to obtain information about asteroids, transform it into a clean format, and save it as a CSV file for further analysis.

The project covers:

Setting up an API key for access to NASA’s API.
Fetching and displaying the Astronomy Picture of the Day (APOD).
Retrieving asteroid data from the NEOWS endpoint.
Preprocessing the asteroid data to obtain a clean dataset.
Exporting the cleaned data to a CSV file.